test(agent): widen the JIT rollout env lock to every test that reads it
`jit_trace_tests` fails intermittently — reproduced 1 run in 10 against a
clean `origin/main` checkout of the djinn-agent lib suite:

    assertion `left == right` failed: unexpected JIT Prometheus delta for eligible_search
      left: 3
     right: 1

`JIT_PITFALLS_ENV_LOCK` was a file-private `Mutex` in `jit_trace_tests.rs`,
scoped — per its own comment — so that "concurrent JIT tests must not observe
each other's env mutation". That scope was too narrow in two ways.

First, `rollout_mode_from_env()` reads `DJINN_JIT_PITFALLS_ROLLOUT` at call
time, and `maybe_append_pitfall_hint` runs on THREE handler paths —
`call_write`, `call_edit`, and `call_apply_patch`. Any concurrently-running
test on any of those paths bumps a counter in the asserted set, corrupting the
before/after delta. This is not limited to tests that pass a `project_id`: a
`None` project records the `error` outcome, and when the gate is off every
call records `disabled_default_off` — both are in `JIT_PITFALL_OUTCOMES`.

Second, the guard was a `Mutex`, so the first panic poisoned it and every
later JIT test panicked on `.lock().unwrap()`. Observed runs failed 5 and 3
tests from a single root failure, which badly obscures triage.

Move the lock to `test_helpers` as an `RwLock` behind two helpers. The JIT
tests mutate the env and take `jit_env_write_guard()`; the 68 write/edit/
apply_patch tests take `jit_env_read_guard()` and therefore still run
concurrently with each other, so the coverage costs no suite wall-clock. Both
helpers use `PoisonError::into_inner` so one failure reports as itself.

`edit_dispatch_tests.rs` was already at 1498 lines against the repository's
1500-line file-size guard, so adding any guard line to it tripped the guard in
CI. Its unicode/CRLF/guard-rejection and apply_patch role-plumbing cases move
to `edit_dispatch_unicode_tests.rs` (1105 + 441 lines), matching how
`process_lease_recovery_tests.rs` was already split for the same reason.

Verified: 25 consecutive full-suite runs clean on the guard approach, against
a baseline that fails about 1 run in 10. An earlier pass that covered only
`call_write` still failed on iteration 5, which is what surfaced the
`call_edit`/`call_apply_patch` paths.

No production code changes; the rollout gate itself is untouched.
